# swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mock API - Treinamentos",
    "version": "1.0.0",
    "description": "API Fake para desenvolvimento do aplicativo de gestão de treinamentos. Construída com json-server e Express."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Servidor local de desenvolvimento" }
  ],
...
  "security": [{"BearerAuth": []}],
  "paths": {
...
    "/api/login": {
      "post": {
        "security": [],
        "tags": ["Autenticação"],
        "summary": "Login",
        "description": "Autentica o usuário e retorna um token simulado. Não valida senha.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "example": { "email": "carlo.souza@empresa.com", "senha": "senhaSegura123" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" },
                "example": {
                  "token": "token-fake",
                  "refreshToken": "refresh-token-fake",
                  "usuario": { "id": 1, "email": "carlo.souza@empresa.com", "funcionarioId": 1, "ativo": true, "criadoEm": "2026-01-10T08:00:00.000Z" }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/funcionarios": {
      "get": {
        "tags": ["Funcionários"],
        "summary": "Listar funcionários",
        "description": "Retorna a lista de funcionários",
        "parameters": [
          { "$ref": "#/components/parameters/Pagina" },
          { "$ref": "#/components/parameters/Limite" },
          { "$ref": "#/components/parameters/Ordenar" },
          { "$ref": "#/components/parameters/Ordem" }
        ],
        "responses": {
          "200": {
            "description": "Lista de funcionários",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Funcionario" } },
                "example": [{ "id": 1, "nome": "Carlos Souza", "matricula": "F001", "cargo": "Analista de Qualidade", "setor": "Qualidade" }]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Funcionários"],
        "summary": "Criar funcionário",
        "description": "Cria um novo funcionário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Funcionario" },
              "example": { "nome": "Maria Oliveira", "matricula": "F002", "cargo": "Técnico de Segurança", "setor": "Operações" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Funcionário criado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Funcionario" },
                "example": { "id": 2, "nome": "Maria Oliveira", "matricula": "F002", "cargo": "Técnico de Segurança", "setor": "Operações" }
              }
            }
          }
        }
      }
    },
...
    "/api/instrutores": {
      "get": {
        "tags": ["Instrutores"],
        "summary": "Listar instrutores",
        "parameters": [{ "$ref": "#/components/parameters/Pagina" }, { "$ref": "#/components/parameters/Limite" }, { "$ref": "#/components/parameters/Ordenar" }, { "$ref": "#/components/parameters/Ordem" }],
        "responses": { "200": { "description": "Lista de instrutores", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Instrutor" } }, "example": [{ "id": 1, "nome": "Marcos Oliveira", "especialidade": "Segurança do Trabalho", "registro": "CREA-SP-123456", "email": "marcos.oliveira@empresa.com", "interno": true }] } } } }
      },
      "post": {
        "tags": ["Instrutores"],
        "summary": "Criar instrutor",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Instrutor" }, "example": { "nome": "João Silva", "especialidade": "Primeiros Socorros", "registro": "CFM-98765", "email": "joao.silva@empresa.com", "interno": false } } } },
        "responses": { "201": { "description": "Instrutor criado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Instrutor" } } } } }
      }
    },
...
    "/api/usuarios": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Listar usuários",
        "parameters": [{ "$ref": "#/components/parameters/Pagina" }, { "$ref": "#/components/parameters/Limite" }, { "$ref": "#/components/parameters/Ordenar" }, { "$ref": "#/components/parameters/Ordem" }],
        "responses": { "200": { "description": "Lista de usuários", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Usuario" } }, "example": [{ "id": 1, "email": "carlo.souza@empresa.com", "funcionarioId": 1, "ativo": true, "criadoEm": "2026-01-10T08:00:00.000Z" }] } } } }
      },
      "post": {
        "tags": ["Usuários"],
        "summary": "Criar usuário",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" }, "example": { "email": "novo@empresa.com", "senha": "senha123", "funcionarioId": 2, "ativo": true, "criadoEm": "2026-08-04T00:00:00.000Z" } } } },
        "responses": { "201": { "description": "Usuário criado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Usuario" } } } } }
      }
    },
...
    "/api/treinamentos": {
      "get": {
        "tags": ["Treinamentos"],
        "summary": "Listar treinamentos",
        "description": "Retorna a lista de treinamentos",
        "parameters": [
          { "$ref": "#/components/parameters/Pagina" },
          { "$ref": "#/components/parameters/Limite" },
          { "$ref": "#/components/parameters/Ordenar" },
          { "$ref": "#/components/parameters/Ordem" },
          { "name": "status", "in": "query", "required": false, "description": "Filtrar por status", "schema": { "type": "string", "enum": ["pendente", "em_andamento", "concluido", "cancelado"] } }
        ],
        "responses": {
          "200": {
            "description": "Lista de treinamentos",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Treinamento" } },
                "example": [{ "id": 1, "titulo": "NR-35 - Trabalho em Altura", "descricao": "Treinamento obrigatório conforme Norma Regulamentadora 35", "cargaHoraria": 8, "status": "concluido", "dataInicio": "2026-03-10T08:00:00.000Z", "dataFim": "2026-03-10T17:00:00.000Z" }]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Treinamentos"],
        "summary": "Criar treinamento",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Treinamento" },
              "example": { "titulo": "NR-10 - Segurança em Instalações Elétricas", "descricao": "Treinamento conforme NR-10", "cargaHoraria": 16, "status": "pendente", "dataInicio": "2026-09-01T08:00:00.000Z", "dataFim": "2026-09-02T17:00:00.000Z" }
            }
          }
        },
        "responses": { "201": { "description": "Treinamento criado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Treinamento" } } } } }
      }
    },
...
    "/api/treinamentos/{id}/completo": {
      "get": {
        "tags": ["Treinamentos"],
        "summary": "Buscar treinamento completo",
        "description": "Retorna o treinamento com instrutores, responsáveis, participantes e evidências",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Treinamento completo",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TreinamentoCompleto" },
                "example": {
                  "treinamento": { "id": 1, "titulo": "NR-35 - Trabalho em Altura", "descricao": "Treinamento obrigatório conforme Norma Regulamentadora 35", "cargaHoraria": 8, "status": "concluido", "dataInicio": "2026-03-10T08:00:00.000Z", "dataFim": "2026-03-10T17:00:00.000Z" },
                  "instrutores": [{ "id": 1, "nome": "Marcos Oliveira", "especialidade": "Segurança do Trabalho", "registro": "CREA-SP-123456", "email": "marcos.oliveira@empresa.com", "interno": true }],
                  "responsaveis": [{ "id": 1, "email": "carlo.souza@empresa.com", "funcionarioId": 1, "ativo": true, "criadoEm": "2026-01-10T08:00:00.000Z" }],
                  "participantes": [{ "id": 1, "treinamentoId": 1, "funcionarioId": 1, "status": "aprovado", "inscritoEm": "2026-03-01T09:00:00.000Z" }],
                  "evidencias": [{ "id": 1, "treinamentoId": 1, "tipo": "foto", "descricao": "Foto do participante durante a prática em altura", "arquivo": "evidencia_carlos_souza_nr35_2026.jpg", "registradoEm": "2026-03-10T14:00:00.000Z" }]
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NaoEncontrado" }
        }
      }
    },
...
    "/api/certificados": {
      "get": {
        "tags": ["Certificados"],
        "summary": "Listar certificados",
        "parameters": [
          { "$ref": "#/components/parameters/Pagina" }, { "$ref": "#/components/parameters/Limite" }, { "$ref": "#/components/parameters/Ordenar" }, { "$ref": "#/components/parameters/Ordem" },
          { "name": "status", "in": "query", "required": false, "description": "Filtrar por status", "schema": { "type": "string", "enum": ["valido", "expirado", "cancelado"] } }
        ],
        "responses": { "200": { "description": "Lista de certificados", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Certificado" } }, "example": [{ "id": 1, "treinamentoParticipantesId": 1, "numero": "CERT-2026-0001", "dataEmissao": "2026-03-11T08:00:00.000Z", "dataValidade": "2027-03-11T08:00:00.000Z", "status": "valido" }] } } } }
      },
      "post": {
        "tags": ["Certificados"],
        "summary": "Emitir certificado",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Certificado" }, "example": { "treinamentoParticipantesId": 2, "numero": "CERT-2026-0002", "dataEmissao": "2026-08-04T08:00:00.000Z", "dataValidade": "2027-08-04T08:00:00.000Z", "status": "valido" } } } },
        "responses": { "201": { "description": "Certificado emitido", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Certificado" } } } } }
      }
    },
...
    "/api/certificados/{id}/completo": {
      "get": {
        "tags": ["Certificados"],
        "summary": "Buscar certificado completo",
        "description": "Retorna o certificado com participante, funcionário e treinamento",
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "Certificado completo",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CertificadoCompleto" },
                "example": {
                  "certificado": { "id": 1, "treinamentoParticipantesId": 1, "numero": "CERT-2026-0001", "dataEmissao": "2026-03-11T08:00:00.000Z", "dataValidade": "2027-03-11T08:00:00.000Z", "status": "valido" },
                  "participante": { "id": 1, "treinamentoId": 1, "funcionarioId": 1, "status": "aprovado", "inscritoEm": "2026-03-01T09:00:00.000Z" },
                  "funcionario": { "id": 1, "nome": "Carlos Souza", "matricula": "F001", "cargo": "Analista de Qualidade", "setor": "Qualidade" },
                  "treinamento": { "id": 1, "titulo": "NR-35 - Trabalho em Altura", "descricao": "Treinamento obrigatório conforme Norma Regulamentadora 35", "cargaHoraria": 8, "status": "concluido", "dataInicio": "2026-03-10T08:00:00.000Z", "dataFim": "2026-03-10T17:00:00.000Z" }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NaoEncontrado" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Identificador único do recurso",
        "schema": { "type": "integer", "example": 1 }
      },
      "Pagina": {
        "name": "_page",
        "in": "query",
        "required": false,
        "description": "Número da página",
        "schema": { "type": "integer", "default": 1, "example": 1 }
      },
      "Limite": {
        "name": "_limit",
        "in": "query",
        "required": false,
        "description": "Quantidade de itens por página",
        "schema": { "type": "integer", "default": 10, "example": 10 }
      },
      "Ordenar": {
        "name": "_sort",
        "in": "query",
        "required": false,
        "description": "Campo para ordenação",
        "schema": { "type": "string", "example": "id" }
      },
      "Ordem": {
        "name": "_order",
        "in": "query",
        "required": false,
        "description": "Direção da ordenação",
        "schema": { "type": "string", "enum": ["asc", "desc"], "default": "asc" }
      }
    },
    "responses": {
...
      "NaoEncontrado": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErroResposta" },
            "example": { "success": false, "error": { "code": "NOT_FOUND", "message": "Registro não encontrado." } }
          }
        }
      },
...
    },
    "schemas": {
      "Funcionario": {
        "type": "object",
        "properties": {
          "id":       { "type": "integer",  "example": 1 },
          "nome":     { "type": "string",   "example": "Carlos Souza" },
          "matricula":{ "type": "string",   "example": "F001" },
          "cargo":    { "type": "string",   "example": "Analista de Qualidade" },
          "setor":    { "type": "string",   "example": "Qualidade" }
        }
      },
      "Instrutor": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer", "example": 1 },
          "nome":         { "type": "string",  "example": "Marcos Oliveira" },
          "especialidade":{ "type": "string",  "example": "Segurança do Trabalho" },
          "registro":     { "type": "string",  "example": "CREA-SP-123456" },
          "email":        { "type": "string",  "format": "email", "example": "marcos.oliveira@empresa.com" },
          "interno":      { "type": "boolean", "example": true }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id":           { "type": "integer", "example": 1 },
          "email":        { "type": "string",  "format": "email", "example": "carlo.souza@empresa.com" },
          "funcionarioId":{ "type": "integer", "example": 1 },
          "ativo":        { "type": "boolean", "example": true },
          "criadoEm":     { "type": "string",  "format": "date-time", "example": "2026-01-10T08:00:00.000Z" }
        }
      },
...
      "Treinamento": {
        "type": "object",
        "properties": {
          "id":         { "type": "integer", "example": 1 },
          "titulo":     { "type": "string",  "example": "NR-35 - Trabalho em Altura" },
          "descricao":  { "type": "string",  "example": "Treinamento obrigatório conforme Norma Regulamentadora 35" },
          "cargaHoraria":{ "type": "integer","example": 8 },
          "status":     { "type": "string",  "enum": ["pendente","em_andamento","concluido","cancelado"], "example": "concluido" },
          "dataInicio": { "type": "string",  "format": "date-time", "example": "2026-03-10T08:00:00.000Z" },
          "dataFim":    { "type": "string",  "format": "date-time", "example": "2026-03-10T17:00:00.000Z" }
        }
      },
...
      "TreinamentoParticipante": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer", "example": 1 },
          "treinamentoId":{ "type": "integer","example": 1 },
          "funcionarioId":{ "type": "integer","example": 1 },
          "status":      { "type": "string",  "enum": ["pendente","aprovado","reprovado"], "example": "aprovado" },
          "inscritoEm":  { "type": "string",  "format": "date-time", "example": "2026-03-01T09:00:00.000Z" }
        }
      },
...
      "Evidencia": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer", "example": 1 },
          "treinamentoId":{ "type": "integer","example": 1 },
          "tipo":        { "type": "string",  "example": "foto" },
          "descricao":   { "type": "string",  "example": "Foto do participante durante a prática em altura" },
          "arquivo":     { "type": "string",  "example": "evidencia_carlos_souza_nr35_2026.jpg" },
          "registradoEm":{ "type": "string",  "format": "date-time", "example": "2026-03-10T14:00:00.000Z" }
        }
      },
      "Certificado": {
        "type": "object",
        "properties": {
          "id":                       { "type": "integer", "example": 1 },
          "treinamentoParticipantesId":{ "type": "integer","example": 1 },
          "numero":                   { "type": "string",  "example": "CERT-2026-0001" },
          "dataEmissao":              { "type": "string",  "format": "date-time", "example": "2026-03-11T08:00:00.000Z" },
          "dataValidade":             { "type": "string",  "format": "date-time", "example": "2027-03-11T08:00:00.000Z" },
          "status":                   { "type": "string",  "enum": ["valido","expirado","cancelado"], "example": "valido" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "senha"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "carlo.souza@empresa.com" },
          "senha": { "type": "string", "example": "senhaSegura123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token":        { "type": "string", "example": "token-fake" },
          "refreshToken": { "type": "string", "example": "refresh-token-fake" },
          "usuario":      { "$ref": "#/components/schemas/Usuario" }
        }
      },
...
      "TreinamentoCompleto": {
        "type": "object",
        "properties": {
          "treinamento": { "$ref": "#/components/schemas/Treinamento" },
          "instrutores": { "type": "array", "items": { "$ref": "#/components/schemas/Instrutor" } },
          "responsaveis":{ "type": "array", "items": { "$ref": "#/components/schemas/Usuario" } },
          "participantes":{ "type": "array", "items": { "$ref": "#/components/schemas/TreinamentoParticipante" } },
          "evidencias":  { "type": "array", "items": { "$ref": "#/components/schemas/Evidencia" } }
        }
      },
      "CertificadoCompleto": {
        "type": "object",
        "properties": {
          "certificado": { "$ref": "#/components/schemas/Certificado" },
          "participante":{ "$ref": "#/components/schemas/TreinamentoParticipante" },
          "funcionario": { "$ref": "#/components/schemas/Funcionario" },
          "treinamento": { "$ref": "#/components/schemas/Treinamento" }
        }
      },
...
      "ErroResposta": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code":    { "type": "string", "example": "NOT_FOUND" },
              "message": { "type": "string", "example": "Registro não encontrado." }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token fake de autenticação. Use o valor: token-fake"
      }
    }
  }
}